requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pandas==2.2.3
pycryptodome==3.21.0
python-dotenv==1.0.1
openpyxl==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
